refactor: Btn and FieldEditor
