+++ web-manager/src/app/components/Btn/CloseButton.tsx
