+++ web-manager/src/app/components/FieldEditor/FieldEditorString.tsx
@@ -1,30 +1,34 @@
-'user client';
+'use client';
 
-import { IJobEntity } from "@/types/IJobEntity";
 import { useEffect, useRef, useState } from "react";
-import BtnSave from "./BtnSave";
 import FieldEditorErrorPanel from "./FieldEditorErrorPanel";
+import BtnSave from "../Btn/BtnSave";
 
-interface FieldEditorDescriptionProps {
-  job: IJobEntity;
+interface FieldEditorStringProps {
+  initialValue: string | null | undefined;
+  legendValue?: string;
   isEditMode: boolean;
-  saveAction?: (value: string | null) => Promise<void>;
+  saveFunction?: (value: string | null) => Promise<void>;
 }
 
-export default function FieldEditorDescription({ job, isEditMode, saveAction }: FieldEditorDescriptionProps) {
+export default function FieldEditorString({ initialValue, legendValue, isEditMode, saveFunction }: FieldEditorStringProps) {
   const ref = useRef<HTMLDivElement>(null);
 
-  const [description, setDescription] = useState<string | null>(job.description || null);
+  const [inputValue, setInputValue] = useState<string | null | undefined>(initialValue);
   const [isEditing, setIsEditing] = useState(false);
   const [error, setError] = useState<string | null>(null);
 
+  const handleremoveError = () => {
+    setError(null);
+  }
+
   const save = async () => {
-    if (saveAction) {
+    if (saveFunction && isEditMode && inputValue !== undefined) {
       try {
-        await saveAction(description || null);
+        await saveFunction(inputValue);
         setIsEditing(false);
       } catch (error) {
-        let errorMessage = "An error occurred while saving the description.";
+        let errorMessage = "An error occurred while saving the value.";
         if (error instanceof Error) errorMessage = error.message;
         else if (typeof error === "string") errorMessage = error;
         setError(errorMessage);
@@ -34,13 +38,9 @@ export default function FieldEditorDescription({ job, isEditMode, saveAction }: 
     }
   }
 
-  const handleChange = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
-    setDescription(e.target.value);
+  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    setInputValue(e.target.value);
   };
-
-  const handleremoveError = () => {
-    setError(null);
-  }
 
   useEffect(() => {
     const handleKeyDown = (e: KeyboardEvent) => {
@@ -66,11 +66,12 @@ export default function FieldEditorDescription({ job, isEditMode, saveAction }: 
   if (isEditMode && isEditing) {
     return (
       <div ref={ref} className="flex items-center bg-blue-100 shadow-md rounded py-1 px-1">
-        <textarea
-          value={description || ''}
+        <input
+          type="text"
+          value={inputValue || ''}
           onChange={handleChange}
-          className="border rounded px-2 py-1 w-full resize-y min-h-[100px] mr-1"
-          placeholder="Description of the job"
+          className="border rounded px-2 py-1 w-full mr-1"
+          placeholder={legendValue ? legendValue : "Enter value"}
         />
         <BtnSave onClick={save} />
         <FieldEditorErrorPanel message={error} close={handleremoveError} />
@@ -81,11 +82,12 @@ export default function FieldEditorDescription({ job, isEditMode, saveAction }: 
   let className = '';
   if (isEditMode) className += 'cursor-pointer hover:text-blue-600';
   return (
-    <span
+    <div
+      ref={ref}
       className={className}
       onClick={() => { if (isEditMode) setIsEditing(true); }}
     >
-      {description || '[N/A]'}
-    </span>
+      {inputValue || '[N/A]'}
+    </div>
   );
 }
